requirements.txt
wheel==0.37.1
Babel==2.9.1
click==8.0.1
colorama==0.4.4
cryptography==36.0.1
cuid==0.3
dnspython==1.15.0
email-validator==1.1.3
# Newer eventlet versions are currently causing conflicts and errors (like version "0.33.1")
eventlet==0.30.2
firebase-admin==6.0.1
Flask==2.0.1
Flask-Babel==2.0.0
Flask-Cors==3.0.10
Flask-Limiter==1.4
flask-marshmallow==0.14.0
Flask-SocketIO==5.3.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gevent-websocket==0.10.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
limits==1.5.1
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.1
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-socketio==5.7.2
pytz==2021.3
scipy==1.8.0
shortuuid==1.0.11
six==1.16.0
SQLAlchemy==1.4.23
typing-extensions==3.10.0.2
urllib3==1.26.7
uuid==1.30
waitress==2.0.0
Werkzeug==2.0.1
WTForms==2.3.3
WTForms-JSON==0.3.3
zipp==3.5.1
